// src/app/core/user/user.types.ts
export interface User
{
    id: string;
    username: string;
    name: string;
    email: string;
    mobile: string;
    whatsapp: string;
    cnic: string;
    avatar?: string;
    candidateid?: number;
    candidate?: string;
    candidatedetail?: string;
    candidatedetailarea?: string;
    party?: string;
    symbol?: string;
    level?: string;
    constituency?: string;
    status?: string;
}